Introduce bugzilla metric and refactor argparser and logging
Add bugzilla modules and integrate them into the project. This comes with also
additional changes in the code. First is the introduction of mongo
db. Bugzilla metrics are making use of nosql with the intention that the same
logic will replace the structural format of redmine metrics. Note that redmine
classes are not touched, neither the sqlalchemy part is removed.

In additional, the argparser is refactored to replace `--source` argument as a
subcommand and with each of those to make use their own arguments. Adapting
the bugzilla and mongo approach, keep in mind that also the `--database` is
not needed as parameter.

Finally logging.yaml is added to provide the logging configuration of the project.

Signed-off-by: ybonatakis <ybonatakis@suse.com>